rename and organize components in separate files
