+++ src/lib/Scenes/MyBids/Components/index.tsx
@@ -1,0 +1,3 @@
+export { ActiveLot } from "./ActiveLot"
+export { MyBidsPlaceholder } from "./MyBidsPlaceholder"
+export { RecentlyClosedLot } from "./RecentlyClosedLot"
